requirements.txt
nltk==3.3
six==1.11.0
Django==2.1.2
djangorestframework==3.9.0
django-cors-headers==2.4.0
